fixed issue with recursively finding patient name
the issue was with how go-pkg-xmlx worked. I sent in a pull
request to add the function SelectNodesDirect, that would not
try to find nodes recursively. The issue is there are other name
fields inside the patient xml tag in a CCD, and if the name
field was missing (honestly shouldn't happen in practice), then
it would pick another name (like guardian).