CCD: Support for <translation> codes in Problems
Problem now contains a Code that is the <translation> tag. Additionally,
if we don't already have a displayName for this problem, we use the
displayName set in the translation tag.

I could not find any example CCDs with more than one translation tag, so
I'm not sure how this code behaves in that circumstance if that is
possible.