Changed how we handle Codes
Code now has an originalText, Translations, and Qualifiers.

Previously our medications and problems parser both had special logic to
look for a translation if the Code was missing and to override the Code
with it.

I moved that check into the Code decoding itself, and changed it to just
override the values rather than the entire element (so that we keep any
extra Translations or Qualifiers).

Note that this is technically not correct, but it gives us better
results for the data we've been dealing with, so I kept it.