Ignore certain sections of the ccd if there are null values.
I don't handle nullFlavors at all here. I'm basically just ensuring that
the individual sections have names to them. So for example, if we
were to process the allergies section and there was one entry but
it was missing a name, we won't add that to the list of allergies.

The one problem this ignores is when nullFlavors are something
that is actually worth knowing, like if a patient refuses to
answer a certain question.

For now, this works for how our backend functions and if we needed
to use nullFlavors properly, we will implement it.